Arm.joints returns 3d points
You can use point[::2] to convert back to a 2d space quickly. Or project
them with a projection matrix if you wish.

Also, the gradient descent uses an adaptive alpha now. Basically, the step
size depends on the size of the error. This has proven to be much more
effective.