Finished multithreading AF code.
It's not pretty, and there's a 0-3% difference in the output between
multithreaded and not, but it's not significant enough for me to be
concerned. Also, this code has turned into something I don't really want
to touch anymore. Who knew it would take 5D arrays to do something like
this in Python? I'll implement the actual control output tomorrow.